+++ pkg/backend/livestream/bilibili/space/notif.go
@@ -35,9 +35,9 @@ func (Data CheckSctruct) SendNude() {
 						SetURL("https://www.bilibili.com/video/"+video.Bvid).
 						AddField("Type ", video.VideoType).
 						AddField("Duration ", video.Length).
+						InlineAllFields().
 						AddField("Viwers ", engine.NearestThousandFormat(float64(video.Play))).
 						SetFooter(durafmt.Parse(time.Now().In(loc).Sub(time.Unix(int64(video.Created), 0).In(loc))).LimitFirstN(2).String()+" Ago", config.BiliBiliIMG).
-						InlineAllFields().
 						SetColor(Color).MessageEmbed)
 					if err != nil {
 						log.Error(msg, err)