+++ pkg/backend/livestream/youtube/youtube.go
@@ -26,7 +26,9 @@ var (
 func Start(BotInit *discordgo.Session, cronInit *cron.Cron) {
 	Bot = BotInit
 	cronInit.AddFunc("@every 0h5m0s", CheckYtSchedule)
+	cronInit.AddFunc("@every 2h30m0s", CheckPrivate)
 	log.Info("Enable youtube module")
+	CheckYtSchedule()
 }
 
 func CheckYtSchedule() {